Moved general functions in their activities
